+++ crates/store/src/state/mod.rs
@@ -171,20 +171,17 @@ impl State {
         let latest_block_num = blockchain.chain_tip().unwrap_or(BlockNumber::GENESIS);
 
         #[cfg(feature = "rocksdb")]
-        let account_storage_config = storage_options.account_tree.into();
+        let (account_storage_config, nullifier_storage_config) =
+            (storage_options.account_tree.into(), storage_options.nullifier_tree.into());
         #[cfg(not(feature = "rocksdb"))]
-        let account_storage_config = {
+        let (account_storage_config, nullifier_storage_config) = {
             let _ = &storage_options;
-            ()
+            ((), ())
         };
         let account_storage =
             TreeStorage::create(data_path, &account_storage_config, ACCOUNT_TREE_STORAGE_DIR)?;
         let account_tree = account_storage.load_account_tree(&mut db).await?;
 
-        #[cfg(feature = "rocksdb")]
-        let nullifier_storage_config = storage_options.nullifier_tree.into();
-        #[cfg(not(feature = "rocksdb"))]
-        let nullifier_storage_config = ();
         let nullifier_storage =
             TreeStorage::create(data_path, &nullifier_storage_config, NULLIFIER_TREE_STORAGE_DIR)?;
         let nullifier_tree = nullifier_storage.load_nullifier_tree(&mut db).await?;